Change folder structure and naming, styles are now separated in new folder
