+++ src/components/Loader.js
@@ -1,4 +1,4 @@
-import './styles.scss';
+import '../styles/components/Loader.scss';
 
 const Loader = () => <div className="box"><div className="Loader" /></div>;
