+++ src/containers/FilterModal.js
@@ -3,9 +3,9 @@ import Modal from 'react-modal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFilter, faTimesCircle, faUndoAlt } from '@fortawesome/free-solid-svg-icons';
 import { useSelector, useDispatch } from 'react-redux';
-import { addFilter, removeFilter, resetFilter } from '../../redux/actions';
-import FilterOptions from '../../components/FilterOptions';
-import './styles.scss';
+import { addFilter, removeFilter, resetFilter } from '../redux/actions';
+import FilterOptions from '../components/FilterOptions';
+import '../styles/containers/FilterModal.scss';
 
 const FilterModal = () => {
   const [showModal, setShowModal] = useState(false);
