+++ src/styles/components/Loader.scss
