+++ src/styles/components/SeasonCard.scss
