+++ src/styles/containers/RankingsList.scss
