+++ src/styles/containers/SeasonsLists.scss
